refactor cross remote item methods
