+++ scenes/xremote_scene_save_remote_item.c
@@ -1,5 +1,5 @@
 #include "../xremote.h"
-#include "../models/cross/xremote_remote.h"
+#include "../models/cross/xremote_cross_remote.h"
 
 #include <string.h>
 #include <toolbox/path.h>
